moved out gene filter table and basic info from dataset page
