+++ src/components/DatasetBasicInfo.css
@@ -1,0 +1,16 @@
+.dataset_info_block{
+    margin-top: 2%;
+}
+
+.dataset_subheader{
+    font-weight: bold;
+}
+
+.line_div_category_header_content{
+    margin-top: 0px;
+}
+
+.dataset_subcontent{
+    color: black;
+    margin-bottom: 3%;
+}
