+++ src/components/DatasetPage.js
@@ -1,6 +1,5 @@
 import React, { useEffect, useState, useRef } from 'react';
 import "./DatasetPage.css";
-import { CircularProgress } from '@mui/material';
 
 import axios from 'axios';
 
@@ -34,16 +33,9 @@ import { useAuth0 } from '@auth0/auth0-react';
 import ProductFilter from './ProductFilter';
 import NumberFilter from './NumberFilter';
 import DatasetNameHolder from './DatasetNameHolder';
-
-const user_get_url = `${process.env.REACT_APP_BACKEND_URL}/api/login`;
-
-const selectCompare = [
-  {value: 0, label: 'None'},
-  {value: 1, label: '<'},
-  {value: 2, label: '>'},
-  {value: 3, label: '='},
-  {value: 4, label: 'between'}
-];
+import DatasetBasicInfo from './DatasetBasicInfo';
+import DatasetGenesListTable from './DatasetGenesListTable';
+import filterNumber from './filterNumber';
 
 function DatasetPage() {
   const [dataset, setDataset] = useState({ "name": "None", "gene_ids": "0", "patient_ids": "0" });
@@ -68,14 +60,7 @@ function DatasetPage() {
   const [modified_patients_list_to_update_back, set_modified_patients_list_to_update_back] = useState({});
   const [prev_patients_list_to_undo, set_prev_patients_list_to_undo] = useState({});
   const [gene_information_expanded, setGene_information_expanded] = useState([{'id':0,'gene_id': "ENT"}]);
-  const [gene_list_filtered , set_gene_list_filtered] = useState([{'id':0,'gene_id': "ENT"}]);
-  const [gene_columns, setGene_columns] = useState([{
-    dataField: 'id',
-    text: ''
-  },{
-    dataField: 'gene_id',
-    text: 'gene_id'
-  }]);
+  
   const [together_data_columns, set_together_data_columns] = useState([{
     dataField: 'id',
     text: ''
@@ -86,7 +71,6 @@ function DatasetPage() {
 
   const [collapse_array, set_collapse_array] = useState([])
 
-  const gene_list_node = useRef(null);
   const dataset_matrix_node = useRef(null);
 
   const { user } = useAuth0();
@@ -142,13 +126,7 @@ function DatasetPage() {
   useEffect(() => {
     let object_information = generateGeneObjs(geneIds);
     setGene_information_expanded(object_information);
-    let copy_obj = clone( object_information);
-    set_gene_list_filtered(copy_obj);
   }, [geneIds])
-
-  useEffect(() => {
-    setGene_columns(generateGeneTable(gene_information_expanded));
-  }, [gene_information_expanded]);
 
   useEffect(() => {
     const edit_recs_url = `${process.env.REACT_APP_BACKEND_URL}/api/edits_dataset_user/all`;
@@ -251,177 +229,6 @@ function DatasetPage() {
 
   const navigate = useNavigate();
 
-  const TextFuzzyFilter = (props) => {
-    const [inputStr, setInputStr] =  useState("");
-
-    const propTypes = {
-      column: PropTypes.object.isRequired,
-      onFilter: PropTypes.func.isRequired
-    }
-
-    const reset_list = () => {
-      props.onFilter(
-        {input_string_value: "", colName: props.column.dataField, reset: true}, gene_information_expanded
-      )
-    }
-    
-    const filter = () => {
-      props.onFilter(
-        {input_string_value: inputStr, colName: props.column.dataField, reset: false}, gene_information_expanded
-      );
-    }
-  
-      return (
-            <div id="column_gene_search_input">
-              <div class="input-group" >
-                <div class="form-outline gene_input_txt">
-                  <input type="text" id="form1" class="form-control" placeholder="text" data-cy="genelistinputbox" onChange={(e) => { setInputStr( e.target.value ) }} />
-                </div>
-                <button type="button"  onClick={() => {filter()}} class="btn btn-primary gene_search_btn">
-                  <i id="gene_list_search" class="fas fa-search"></i>
-                </button>
-                <button
-                    id="gene_list_reset"
-                    onClick={() => {reset_list()}}
-                  >Reset</button>
-              </div>
-
-            </div>
-          )
-      }
-    
-    const hasLevenshteinDistanceLessThanEqualOne = (current_input_str, str_reference) => {
-      if( Math.abs(current_input_str.length - str_reference.length) > 1 ){
-        // rand number more than 1 to be discarded when filtered
-        return 5;
-      }else if( Math.abs(current_input_str.length - str_reference.length) == 1 ){
-        // distance is 1, make sure no other diff, or else not dist 1
-
-        let larger_str = current_input_str;
-        let smaller_str = str_reference;
-        if(current_input_str.length < str_reference.length){
-          larger_str = str_reference;
-          smaller_str = current_input_str;
-        }
-
-
-        let num_errors = 0;
-        let smaller_index = 0;
-        let larger_index = 0;
-
-        for( ; smaller_index < smaller_str.length && larger_index < larger_str.length; ){
-          if( smaller_str[smaller_index] == larger_str[larger_index]  ){
-            smaller_index++;
-            larger_index++;
-          } else if( (larger_index+1) < larger_str.length && smaller_str[smaller_index] == larger_str[larger_index+1] ){
-            larger_index = larger_index+2;
-            smaller_index++;
-            num_errors++;
-          } else {
-            larger_index++;
-            smaller_index++;
-            num_errors++;
-          }
-        }
-
-        
-        return num_errors;
-
-      } else {
-        // equal length
-
-        let num_errors = 0;
-        let first_index = 0;
-        let second_index = 0;
-
-        for( ; first_index < current_input_str.length && second_index < str_reference.length; ){
-          if( current_input_str[first_index] == str_reference[second_index]  ){
-            first_index++;
-            second_index++;
-          } else {
-            second_index++;
-            first_index++;
-            num_errors++;
-          }
-        }
-
-      
-        return num_errors;
-      }
-      
-    }
-
-    const filterFuzzyText = (filterVals, data) => {
-      let input_str = filterVals['input_string_value']
-      let colName = filterVals['colName']
-
-      if( filterVals['reset'] == true ){
-        return data;
-      }
-
-      // not exact match or exact off by 1: need to be ok with includes
-
-      // also discarding decimal in search important
-      let removed_decimal_input = input_str
-      if( removed_decimal_input.includes(".") ){
-        removed_decimal_input = removed_decimal_input.substring(0, removed_decimal_input.indexOf("."))
-      }
-      
-      // equals filter
-      let filtered_list_genes = data.filter( patient_one => (hasLevenshteinDistanceLessThanEqualOne(patient_one[colName] , input_str) <= 1) || (patient_one[colName].includes(input_str)) || (patient_one[colName].includes(removed_decimal_input)) ).sort(
-        (patient_object_a, patient_object_b) => {
-          return hasLevenshteinDistanceLessThanEqualOne(patient_object_a[colName] , input_str) - hasLevenshteinDistanceLessThanEqualOne(patient_object_b[colName] , input_str);
-        }
-      );
-
-      return filtered_list_genes;
-      
-    }
-
-  const filterNumber = (filterVals, data) => {
-    let compareValCode = filterVals['compareValCode']
-    let inputVal1 = filterVals['inputVal1']
-    let inputVal2 = filterVals['inputVal2']
-    let colName = filterVals['colName']
-    
-    if(compareValCode == 0){
-      // no filter
-      return data;
-    }
-    else if(compareValCode == 1){
-      // <
-      return data.filter(patient_one => patient_one[colName] < inputVal1);
-    } else if (compareValCode == 2){
-      // >
-      return data.filter(patient_one => patient_one[colName] > inputVal1);
-    } else if (compareValCode == 3){
-      // =
-      return data.filter(patient_one => patient_one[colName] == inputVal1);
-    } else if(compareValCode == 4){
-      // Between
-      return data.filter(patient_one => patient_one[colName] > inputVal1 && patient_one[colName] < inputVal2 );
-    }
-
-    
-  }
-
-  const geneListFilter = (gene_list_filter_value) => {
-    //console.log("node method");
-    //console.log(gene_list_filter_value)
-
-    let text_search = gene_list_filter_value.gene_id.filterVal.input_string_value ;
-    let column_search = gene_list_filter_value.gene_id.filterVal.colName;
-    let reset_value = gene_list_filter_value.gene_id.filterVal.reset;
-
-    if(reset_value == true){
-      set_gene_list_filtered(clone(gene_information_expanded));
-      return;
-    }
-
-    let search_results_genes = filterFuzzyText({input_string_value: text_search, colName: column_search, reset: false}, gene_information_expanded);
-
-    set_gene_list_filtered(clone(search_results_genes));
-  }
 
   const generateGeneObjs = (gene_ids_info) => {
     if(gene_ids_info == null || gene_ids_info.length == 0){
@@ -440,98 +247,6 @@ function DatasetPage() {
     //console.log(gene_objs)
 
     return gene_objs;
-  }
-
-  const generateGeneTable = (gene_objs_information) => {
-
-    // 'id' not need options
-    let gene_columns_list = []
-
-    if(gene_objs_information == null || gene_objs_information.length == 0){
-      return [{
-        dataField: 'id',
-        text: ''
-      },{
-        dataField: 'gene_id',
-        text: 'gene_id'
-      }];
-    }
-
-    let column_possibilities = ['gene_id']
-    for(let i = 0; i < column_possibilities.length; i++){
-      let unique = [...new Set(gene_objs_information.flatMap(item => item[ column_possibilities[i] ] ))];
-
-      let select_options_col = []
-
-      for(let j = 0; j < unique.length; j++){
-        select_options_col.push({value: unique[j], label: unique[j]})
-      }
-
-      let col_obj = {dataField: column_possibilities[i],
-        text: column_possibilities[i]}
-      if(unique.length > 0 && Number.isInteger(unique[0])){
-        col_obj = {
-          dataField: column_possibilities[i],
-          text: column_possibilities[i],
-          headerStyle: { minWidth: '150px' },
-          filter: customFilter({
-            delay: 1000,
-            onFilter:filterNumber,
-            type: FILTER_TYPES.NUMBER
-          }),
-          filterRenderer: (onFilter, column) => {
-            return(
-              <NumberFilter onFilter={ onFilter } column={column} />
-              )
-          }
-        }
-      }
-      else if(unique.length < 5){
-        col_obj = {
-          dataField: column_possibilities[i],
-          text: column_possibilities[i],
-          headerStyle: { minWidth: '150px' },
-          filter: customFilter({
-            delay: 1000,
-            type: FILTER_TYPES.MULTISELECT
-          }),
-        
-          filterRenderer: (onFilter, column) => {
-            return(
-              <ProductFilter onFilter={onFilter} column={column} optionsInput={clone(select_options_col)}/>
-              )
-          }
-        }
-      } else {
-        col_obj = {
-          dataField: column_possibilities[i],
-          text: column_possibilities[i],
-          headerStyle: { minWidth: '150px' },
-          formatter: (cell, row, rowIndex, extraData) => {
-            return(
-              <span>
-                <a id="gene_list_single_link" href={"/gene/"+ cell +"/1"}>{cell}</a>
-              </span>
-            );
-          },
-          formatExtraData: gene_list_filtered,
-          filter: customFilter({
-            delay: 1000,
-            onFilter:filterFuzzyText
-          }),
-          filterRenderer: (onFilter, column) => {
-            return(
-              <TextFuzzyFilter onFilter={ onFilter } column={column} />
-              )
-          }
-        }
-      }
-      gene_columns_list.push(col_obj)
-    }
-    //console.log("gene info:");
-    //console.log(gene_columns_list);
-    //console.log(gene_information_expanded);
-    return gene_columns_list;
   }
 
   const generateDatasetMatrixTable = () => {
@@ -906,7 +621,7 @@ function DatasetPage() {
 
               </div>
 
-              <DatasetNameHolder input_dataset_id={DATASET_ID} input_dataset={dataset}/>
+              <DatasetNameHolder input_dataset_id={DATASET_ID} input_dataset={dataset} />
 
               <div class="container-fluid" id="tabs_container" >
                   <Tabs
@@ -915,60 +630,10 @@ function DatasetPage() {
                       className="mb-3"
                     >
                       <Tab eventKey="basic_info" title="Basic Info">
-                          <div  id="dataset_basics">
-                          
-                            <div class="card shadow dataset_info_block" >
-                            
-                              <div class="card-body">
-                              
-                                <div >
-                                  <h5 class="dataset_subheader">Description </h5>
-                                  <hr class="line_div_category_header_content" />
-                                  <p class="dataset_subcontent">{dataset["description"]}</p>
-                                </div>
-
-                              
-                              {datasetTableInputFormat.length>3 ? (
-                                <div>
-                                  {datasetTableInputFormat.map((table_property) => {
-
-                                    if(table_property.field_name != 'id' && table_property.field_name != 'name' && table_property.field_name != 'description'){
-                                      //console.log("info dataset table objs: !!!: ")
-                                      //console.log(table_property)
-                                      return(<div >
-                                        <h5 class="dataset_subheader">{table_property.field_name} </h5>
-                                        <hr class="line_div_category_header_content" />
-                                        <p class="dataset_subcontent">{table_property.value}</p>
-                                      </div>)
-                                    }
-                                  })}
-                                </div>
-                                ):(
-                                  <div>
-                                    <CircularProgress />
-                                  </div>
-                                )
-                              }
-                              </div>
-                            </div>
-                            
-                          </div>
+                          <DatasetBasicInfo input_datasetTableInputFormat={datasetTableInputFormat} input_dataset={dataset}/>
                       </Tab>
                       <Tab eventKey="genes_list" title="Genes List">
-                        <div class="row" id="dataset_table_and_stats_row">
-                            <div class="card shadow" id="dataset_genes_list">
-                              <div class="card-header py-3">
-                                <h6 class="m-0 font-weight-bold text-primary">Gene Ids</h6>
-                              </div>
-                              <div class="card-body">
-                                  <div class="row" id="table_options_outer">
-                                      <div id="gene_table_area">
-                                        <BootstrapTable keyField='id' ref={ n => gene_list_node.current = n  } remote={ { filter: true, pagination: false, sort: false, cellEdit: false } } data={ gene_list_filtered } columns={ gene_columns } filter={ filterFactory() } pagination={ paginationFactory() } filterPosition="top" onTableChange={ (type, newState) => { geneListFilter(gene_list_node.current.filterContext.currFilters) } } />
-                                      </div>
-                                  </div>
-                              </div>
-                            </div>
-                          </div>
+                        <DatasetGenesListTable input_expanded_gene_info={gene_information_expanded} />
                       </Tab>
                       <Tab eventKey="edit_dataset" title="Edit Dataset">
                         <div id="dataset_view_table"></div>
