+++ src/components/filterNumber.js
@@ -1,0 +1,30 @@
+
+
+const filterNumber = (filterVals, data) => {
+    let compareValCode = filterVals['compareValCode']
+    let inputVal1 = filterVals['inputVal1']
+    let inputVal2 = filterVals['inputVal2']
+    let colName = filterVals['colName']
+    
+    if(compareValCode == 0){
+      // no filter
+      return data;
+    }
+    else if(compareValCode == 1){
+      // <
+      return data.filter(patient_one => patient_one[colName] < inputVal1);
+    } else if (compareValCode == 2){
+      // >
+      return data.filter(patient_one => patient_one[colName] > inputVal1);
+    } else if (compareValCode == 3){
+      // =
+      return data.filter(patient_one => patient_one[colName] == inputVal1);
+    } else if(compareValCode == 4){
+      // Between
+      return data.filter(patient_one => patient_one[colName] > inputVal1 && patient_one[colName] < inputVal2 );
+    }
+
+    
+  }
+
+export default filterNumber
